requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
python-jose[cryptography]
supabase
pandas
numpy
openpyxl
prophet
plotly
groq
google-genai
cerebras-cloud-sdk
fpdf2
kaleido
python-dotenv
pdfplumber
python-docx
scikit-learn
scipy
pystan
